Tweak: replace 0 counts by small number (hand)
TODO: find routine to determine what that number should be. Right now done by hand.

Tweaked the matrixes, such that we replace places in the matrix with 0 counts by a small number. This improves the fit.

Previous problem with 0 counts: the chi2 function is weighted with the "exp." variance (or rather 1/var). For a poisson this is given by #cnt. If #cnt is 0, the weighting is 1/0. This is math. undefined, and in the code fixed to be 0. Thus the pixel in P_fit can have any value.

If we replace the 0 counts by a small number, then its influence on the chi2 remains.