Started uncertainty estimate on fitted rho and T
Started to estimate uncertainty from the fir to the 1Gen matrix. We perturb the input 1Gen matrix and take the average and stddev of the results.

Note: This works to some extend only, because the minimization routine has "issues". If the minimization would not always run into some sort of /the same/ minima, we would expect to get very different nlds and Ts every time, as we could get any other transformation, too. However, this does not seem to be the case (spread in extracted nld and T is quite small)

TODO: Process the derived uncertainties all the way. For now this is only processed for the normalization of the nld. Not in the normalization of the gSF and not in the plots.

Note: You can get the uncertainties for the nld normalization derived from multinest running `multinest_marginals.py multinest/norm_3_`

Sample output:
```
model "multinest/norm_3_"
  analysing data from multinest/norm_3_.txt
  marginal likelihood:
    ln Z = -40.4 +- 0.0
  parameters:
    A              6.35 +- 0.17
    alpha          2.3871 +- 0.0063
    T              0.4303 +- 0.0011
creating marginal plot ...
```